Merge branch for v1.0.24

+++ d3x-morpheus-core/src/main/java/com/d3x/morpheus/array/dense/DenseArrayConstructor.java
@@ -67,7 +67,7 @@ public class DenseArrayConstructor implements ArrayFactory.Constructor {
 
 
     @Override
-    public <T> Array<T> apply(Class<T> type, int length, double fillPct, T defaultValue) {
+    public <T> Array<T> apply(Class<T> type, int length, float fillPct, T defaultValue) {
         return apply(type, length, defaultValue, null);
     }
